Add new example: app-minify-key
