+++ Cargo.toml
@@ -39,7 +39,7 @@ members = [
     "crates/support",
     "crates/macro",
     "examples/app-load-path",
-    "examples/app-tr",
+    "examples/app-minify-key",
     "examples/foo",
 ]
